Add migrations for Event and EloSnapshot
Optional `start` and `end` timestamps were added to the Event
model. Additionally, a unique partial index was added to
EloSnapshot to ensure that for each player/version, no more
than one EloSnapshot can have the `current` flag set. The
purpose of this is to permit more efficient calculation of
both a player's current Elo and, by extension, the overall
Elo rankings. This is anticipated to be a high-demand query,
perhaps even the *highest* demand query, so indexing these
fields makes sense.