Refactor resolvers to reflect best practices
After reading what I really should've read *before* starting this
project, I've been made aware that GraphQL best practices and
conventions are not in line with the previous direction of the project.
As such, I've reafctored to give Input types to each mutation, and to
ensure that all mutations are accomplishing very specific tasks, rather
than trying to craft flexible mutations that handled complex logic on
the server side. This also allows a return to the functional resolver
paradigm that is strongly preferable to unnecessary object-orientation
and/or overengineering.

See: https://graphql.org/learn/best-practices