Add Elo computed values table
Our solution to the issue of computed values in the context of Elo
is to have a "computed values" *table* that updates with each new match
added to the database alongside existing Elo tables. This allows the
computation to be a part of the match addition and to persist, limiting
resource-intensive calculations or queries, particularly surrounding
something like a player's rank.

This commit adds the computed values table, id'ing and indexing it by
Player *and* EloVerison to enable support for distinct Elo algorithms or
implementations existing and updating concurrently.