Optimize nontrivial field resolvers
Prisma has an interesting n+1 problem for recursive field queries,
which is especially prudent in GraphQL applications where resolvers are
computed in exactly that way.

Often, it is necessary to get all results in a table that has a
many-to-one relation with the base model (ex: Matches on an Event
model). A standard `findMany` call on the Matches table would occur only
after a previous reoslver had queried the Events table.

This particular approach will lead to an n+1 problem when running a
`findMany` query on the Events table and then getting all matches from
each. The solution is to use a second `findUnique` query with their
Fluent API to allow the Prisma dataloader to combine the calls into one
query.

Fluent API:
https://www.prisma.io/docs/concepts/components/prisma-client/relation-queries#fluent-api

More details on the problem can be found here:
https://www.prisma.io/docs/guides/performance-and-optimization/query-optimization-performance#solving-the-n1-problem