Add database integration to Elo module
Due to the manner in which Elo is calculated, the in-memory
representations of matches differ from those of the data on which the
Elo algorithms actually operate. This serialization process is embedded
within the interface between those in-memory representations (specified
in src/elo/util/types.ts) and the database itself.