Add dev repl to npm scripts
This small change allows a development TypeScript REPL to be started
through npm scripts with `npm run repl`. It provides a small amount of
boilierplate that can easily be added to, but due to the finnicky nature
of REPL async/await, that setup cannot be fully automated; a helpful
message about acquiring a GraphQLContext object is provided instead.