Refactor intake
It's working! I refactored a couple things for the sake of clarity and
to make it all a little easy to work with, but `npm run intake` does
that now! I may make further tweaks in the future, since this dataset
only goes through 2021 and we're gathering another spreadsheet for 2022
and the first half of 2023, but for the time being this is working and
all 24,000 matches are visible for me locally!