add test db as docker container
This summary was added retroactively; as such, I'm not completely sure
about the justification for each change.

In essence, the testing strategy spins up a brand new postgres database,
runs migrations, then the actual test commands, and finally destroys the
test database. If my memory is correct, we map the test container's port 5432 to 5433 to distinguish between the URLs of the test and dev database.

I also changed the main server code to listen to all interfaces, so that
it could connect to the test database in docker. This should probably be
changed eventually.