rename playstyle field in results
Slight API change: all other types (e.g. Player) refer to this
information as "playstyle," so "style" was renamed to be consistent.

The corresponding migration was manually edited to rename the column, since the
default behavior is to delete the existing column/replace it with a new
one, which wasn't necessary here.