Move test script to separate file
This commit also removes unnecessary print statements in the tests.

I originally needed a `sleep 1` command to wait for the container to
become ready, but I tried removing it and I haven't had any issues, so I
will leave this out for now and hope that the problems don't resurface.