check for passphrase header before executing query
Figured it was sensible to keep the phrase as an environment variable.
This implementation just checks if the proper header is set. If no
header is provided or the value is incorrect, an appropriate graphql
error is sent.

When we move to a more robust authentication system, I think we will
want to move this logic into the contextFactory, which is recommended by
https://graphql-helix.vercel.app/docs/recipes/authentication-and-authorization.
However, right now I think this requires manually handling permissions
for every single query, so I think this should be saved for post-MVP.