allow passphrase disabling for development
motivated because graphiql stopped loading our schema because the
request it sent didn't include the passphrase lmao. having either an
empty string for PASSPHRASE or simply not including the field will
disable the check.

i think this motivates the creation of a .env.development file, so dev
and prod builds can handle this differently, but it should be fine for
now.